device/requirements.txt
# Dagr - Python Dependencies
# FastAPI web application for API token management and display control

# Web framework and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6

# HTTP client for external API requests
requests>=2.31.0

# Security and encryption
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0

# Data validation and serialization
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Configuration and data handling
pyyaml>=6.0
jsonschema>=4.17.0

# Logging and monitoring
psutil>=5.9.0

# Image processing (for display functionality)
Pillow>=10.0.0
numpy>=1.24.0

# E-ink display support (Pimoroni Inky)
inky>=2.2.1

# Scheduling and task management
schedule>=1.2.0
python-crontab>=2.7.0

# Version management and updates
packaging>=23.0
